+++ db/migrate/20201125210614_create_transactions.rb
@@ -1,7 +1,7 @@
 class CreateTransactions < ActiveRecord::Migration[6.0]
   def change
     create_table :transactions do |t|
-      t.integer :size
+      t.float :size
       t.string :side
       t.belongs_to :crypto, null: false, foreign_key: true
       t.belongs_to :user, null: false, foreign_key: true
